+++ resources/js/common/types/globals.d.ts
@@ -266,11 +266,7 @@ interface CraftStatic {
     assetId: number,
     settings?: LegacyWidgetSettings
   ) => LegacyModalInstance;
-  createUploader(
-    fsType: string,
-    $element: JQuery,
-    settings?: LegacyWidgetSettings
-  ): {setParams(params: LegacyWidgetSettings): void};
+  Uploaders: typeof import('@/modules/uploader');
 }
 
 // oxlint-disable-next-line @typescript-eslint/no-empty-object-type