+++ src/Asset/Elements/Asset.php
@@ -1476,20 +1476,18 @@ JS, [
                 'showInChips' => false,
             ];
 
-            HtmlStack::jsWithVars(fn ($id, $namespace, $assetId, $fsType, $dimensionsLabel) => <<<JS
+            HtmlStack::jsWithVars(fn ($id, $namespace, $assetId, $dimensionsLabel) => <<<JS
 $('#' + $id).on('activate', () => {
   const fileInput = $('<input/>', {type: 'file', name: 'replaceFile', class: 'replaceFile hidden'}).appendTo(Garnish.\$bod);
-  const uploader = Craft.createUploader($fsType, fileInput, {
+  const uploader = new Craft.Uploaders.Uploader(fileInput[0], {
     dropZone: null,
-    fileInput: fileInput,
-    paramName: 'replaceFile',
+    fileInput: fileInput[0],
     replace: true,
-    events: {
-      fileuploadstart: () => {
+    on: {
+      start: () => {
         $('#' + Craft.namespaceId('thumb-container', $namespace)).addClass('loading');
       },
-      fileuploaddone: (event, data) => {
-        const result = event instanceof CustomEvent ? event.detail : data.result;
+      done: ({result}) => {
 
         // Update the filename input and serialized param value
         const filenameInput = $('#' + Craft.namespaceId('new-filename', $namespace))
@@ -1563,15 +1561,14 @@ $('#' + $id).on('activate', () => {
           }
         }
       },
-      fileuploadfail: (event, data) => {
-        const file = data.data.getAll('replaceFile');
-        const backupFilename = file[0].name;
+      fail: ({error, canceled, file}) => {
+        if (canceled) {
+          return;
+        }
 
-        const response = event instanceof Event
-          ? event.detail
-          : data?.jqXHR?.responseJSON;
-
-        let {message, filename} = response || {};
+        const backupFilename = file?.name;
+        let message = error instanceof Error ? error.message : undefined;
+        let filename = error?.data?.filename;
 
         if (!message) {
           if (!filename) {
@@ -1584,7 +1581,7 @@ $('#' + $id).on('activate', () => {
 
         Craft.cp.displayError(message);
       },
-      fileuploadalways: (event, data) => {
+      settled: () => {
         $('#' + Craft.namespaceId('thumb-container', $namespace)).removeClass('loading');
       },
     }
@@ -1600,7 +1597,6 @@ JS, [
                 InputNamespace::namespaceId($replaceId),
                 InputNamespace::get(),
                 $this->id,
-                $this->getVolume()->sourceFilesystemType(),
                 t('Dimensions'),
             ]);
         }